Move Azure Functions to subfolder
