--- apps/sending-func/src/GetRCConfiguration/handler.ts
@@ -1,238 +1,0 @@
-import {
-  RCConfigurationModel,
-  RetrievedRCConfiguration,
-} from "@pagopa/io-functions-commons/dist/src/models/rc_configuration";
-import { ContextMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/context_middleware";
-import { RequiredParamMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/required_param";
-import { retrievedRCConfigurationToPublic } from "@pagopa/io-functions-commons/dist/src/utils/rc_configuration";
-import {
-  withRequestMiddlewares,
-  wrapRequestHandler,
-} from "@pagopa/io-functions-commons/dist/src/utils/request_middleware";
-import {
-  IResponseErrorForbiddenNotAuthorized,
-  IResponseErrorInternal,
-  IResponseErrorNotFound,
-  IResponseSuccessJson,
-  ResponseErrorForbiddenNotAuthorized,
-  ResponseErrorInternal,
-  ResponseErrorNotFound,
-  ResponseSuccessJson,
-} from "@pagopa/ts-commons/lib/responses";
-import { NonEmptyString, Ulid } from "@pagopa/ts-commons/lib/strings";
-import * as express from "express";
-import * as E from "fp-ts/lib/Either";
-import { Json } from "fp-ts/lib/Json";
-import { parse } from "fp-ts/lib/Json";
-import * as O from "fp-ts/lib/Option";
-import * as TE from "fp-ts/lib/TaskEither";
-import { flow, pipe } from "fp-ts/lib/function";
-
-import { RCConfigurationResponse } from "../generated/definitions/RCConfigurationResponse";
-import {
-  RequiredSubscriptionIdMiddleware,
-  RequiredUserGroupsMiddleware,
-  RequiredUserIdMiddleware,
-} from "../middlewares/required_headers_middleware";
-import { IConfig } from "../utils/config";
-import { RedisClientFactory } from "../utils/redis";
-import { getTask, setWithExpirationTask } from "../utils/redis_storage";
-import { checkGroupAndManageSubscription } from "../utils/remote_content";
-
-export const RC_CONFIGURATION_REDIS_PREFIX = "RC-CONFIGURATION";
-
-interface IHandlerParameter {
-  readonly configurationId: Ulid;
-  readonly subscriptionId: NonEmptyString;
-  readonly userGroups: NonEmptyString;
-  readonly userId: NonEmptyString;
-}
-
-interface IGetRCConfigurationHandlerParameter {
-  readonly config: IConfig;
-  readonly rccModel: RCConfigurationModel;
-  readonly redisClient: RedisClientFactory;
-}
-
-export const handleErrorResponse = (error: Error): IResponseErrorInternal =>
-  ResponseErrorInternal(
-    `Something went wrong trying to retrieve the configuration: ${error}`,
-  );
-
-export const handleEmptyErrorResponse =
-  (configurationId: Ulid) =>
-  (
-    maybeRCConfiguration: O.Option<RetrievedRCConfiguration>,
-  ): TE.TaskEither<IResponseErrorNotFound, RetrievedRCConfiguration> =>
-    pipe(
-      maybeRCConfiguration,
-      TE.fromOption(() =>
-        ResponseErrorNotFound(
-          `Configuration not found`,
-          `Cannot find any configuration with configurationId: ${configurationId}`,
-        ),
-      ),
-    );
-
-const parseRedisValue: (redisValue: string) => E.Either<Error, Json> = flow(
-  parse,
-  E.mapLeft(() => new Error("Cannot parse RCConfiguration Json from Redis")),
-);
-
-const validateRedisValue: (
-  redisValue: Json,
-) => E.Either<Error, RetrievedRCConfiguration> = flow(
-  RetrievedRCConfiguration.decode,
-  E.mapLeft(() => new Error("Cannot decode RCConfiguration Json from Redis")),
-);
-
-const parseAndValidateRedisValue: (
-  redisValue: string,
-) => E.Either<Error, RetrievedRCConfiguration> = flow(
-  parseRedisValue,
-  E.chain(validateRedisValue),
-);
-
-const handleRedisEmptyResponse: (
-  maybeRedisResponse: O.Option<string>,
-) => TE.TaskEither<Error, string> = TE.fromOption(
-  () => new Error("Cannot Get RCConfiguration from Redis"),
-);
-
-/**
- * This method is used to get a configuration, if it exists, by configuratin id.
- * If the configuration is cached it retrieves it from cache, or else it retrieves
- * it from cosmosdb and provides to cache it.
- *
- * @param configurationId
- * @returns
- */
-const getOrCacheMaybeRCConfigurationById = (
-  configurationId: Ulid,
-  redisClient: RedisClientFactory,
-  rccModel: RCConfigurationModel,
-  config: IConfig,
-): TE.TaskEither<Error, O.Option<RetrievedRCConfiguration>> =>
-  pipe(
-    getTask(redisClient, `${RC_CONFIGURATION_REDIS_PREFIX}-${configurationId}`),
-    TE.chain(handleRedisEmptyResponse),
-    TE.chainEitherK(parseAndValidateRedisValue),
-    TE.fold(
-      () =>
-        pipe(
-          rccModel.findByConfigurationId(configurationId),
-          TE.mapLeft(
-            (e) =>
-              new Error(`${e.kind}, RCConfiguration Id=${configurationId}`),
-          ),
-          TE.chainFirstW((maybeRCConfiguration) =>
-            pipe(
-              maybeRCConfiguration,
-              TE.fromOption(
-                () =>
-                  new Error(
-                    `Cannot find any configuration with id: ${configurationId}`,
-                  ),
-              ),
-              TE.chain((rCConfiguration) =>
-                setWithExpirationTask(
-                  redisClient,
-                  `${RC_CONFIGURATION_REDIS_PREFIX}-${configurationId}`,
-                  JSON.stringify(rCConfiguration),
-                  config.RC_CONFIGURATION_CACHE_TTL,
-                ),
-              ),
-              TE.orElseW(() => TE.of(maybeRCConfiguration)),
-            ),
-          ),
-        ),
-      (rCConfiguration) => TE.right(O.some(rCConfiguration)),
-    ),
-  );
-
-export const getRCConfigurationHandler =
-  ({ config, rccModel, redisClient }: IGetRCConfigurationHandlerParameter) =>
-  ({
-    configurationId,
-    subscriptionId,
-    userGroups,
-    userId,
-  }: IHandlerParameter): Promise<
-    | IResponseErrorForbiddenNotAuthorized
-    | IResponseErrorInternal
-    | IResponseErrorNotFound
-    | IResponseSuccessJson<RCConfigurationResponse>
-  > =>
-    pipe(
-      config.INTERNAL_USER_ID === userId
-        ? TE.of(true)
-        : checkGroupAndManageSubscription(subscriptionId, userGroups),
-      TE.chainW(() =>
-        pipe(
-          getOrCacheMaybeRCConfigurationById(
-            configurationId,
-            redisClient,
-            rccModel,
-            config,
-          ),
-          TE.mapLeft(handleErrorResponse),
-        ),
-      ),
-      TE.chainW(handleEmptyErrorResponse(configurationId)),
-      TE.chainW(
-        TE.fromPredicate(
-          (retrievedConfiguration) =>
-            config.INTERNAL_USER_ID === userId ||
-            retrievedConfiguration.userId === userId,
-          () => ResponseErrorForbiddenNotAuthorized,
-        ),
-      ),
-      TE.map((retrievedConfiguration) =>
-        pipe(
-          retrievedConfiguration,
-          retrievedRCConfigurationToPublic,
-          (publicConfiguration) => ({
-            ...publicConfiguration,
-            user_id: retrievedConfiguration.userId,
-          }),
-          ResponseSuccessJson,
-        ),
-      ),
-      TE.toUnion,
-    )();
-
-interface IGetGetRCConfigurationHandlerParameter {
-  readonly config: IConfig;
-  readonly rccModel: RCConfigurationModel;
-  readonly redisClient: RedisClientFactory;
-}
-
-type GetGetRCConfigurationHandlerReturnType = express.RequestHandler;
-
-type GetGetRCConfigurationHandler = (
-  parameter: IGetGetRCConfigurationHandlerParameter,
-) => GetGetRCConfigurationHandlerReturnType;
-
-export const getGetRCConfigurationExpressHandler: GetGetRCConfigurationHandler =
-  ({ config, rccModel, redisClient }) => {
-    const handler = getRCConfigurationHandler({
-      config,
-      rccModel,
-      redisClient,
-    });
-
-    const middlewaresWrap = withRequestMiddlewares(
-      ContextMiddleware(),
-      RequiredSubscriptionIdMiddleware(),
-      RequiredUserGroupsMiddleware(),
-      RequiredUserIdMiddleware(),
-      RequiredParamMiddleware("configurationId", Ulid),
-    );
-
-    return wrapRequestHandler(
-      middlewaresWrap(
-        (_, subscriptionId, userGroups, userId, configurationId) =>
-          handler({ configurationId, subscriptionId, userGroups, userId }),
-      ),
-    );
-  };
