+++ apps/sending-func/src/functions/CreateRCConfiguration/function.json
