+++ apps/sending-func/src/functions/GetRCConfiguration/function.json
