+++ apps/sending-func/src/functions/ListRCConfiguration/function.json
