+++ apps/sending-func/src/functions/UpdateRCConfiguration/function.json
